requirements_deploy.txt
streamlit>=1.26.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.17.0
matplotlib>=3.7.0
Pillow>=10.0.0